requirements.txt
asgiref==3.11.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.9
django-allauth==65.13.1
django-otp==1.6.3
django-simple-history==3.10.1
djangorestframework==3.16.1
gunicorn==23.0.0
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
qrcode==8.2
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.4
whitenoise==6.11.0
